server/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
django-cors-headers==4.6.0
django-cors-middleware==1.5.0
django-ninja==1.3.0
idna==3.10
psycopg==3.2.3
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
